Simplify autoextract structure
